% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{adjacency}
\alias{adjacency}
\title{River network adjacency matrix}
\usage{
adjacency(x, weighted = FALSE)
}
\arguments{
\item{x}{A river network}

\item{weighted}{boolean, default to FALSE, if true returns the weighted adjacency matrix}
}
\value{
A matrix Y, such that non-zero values in Y\link{i,j} indicate flow from i to j
}
\description{
River network adjacency matrix
}
