% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_funs.R
\name{ce_funs}
\alias{ce_funs}
\alias{ce_linear}
\alias{ce_constant}
\alias{ce_gaussian}
\title{Colonisation and extinction functions}
\usage{
ce_linear(parm)

ce_constant(parm)

ce_gaussian(parm)
}
\arguments{
\item{parm}{Parameter list, a named list containing parameters for the c/e functions; see 'details'}

\item{x}{state variable matrix; input for c/e functions; one row per site, one column per variable}
}
\value{
For c/e functions, a vector of colonisation/extinction rates with length == \code{nrow(x)}
For all others, a c/e function of the desired form
}
\description{
Colonisation and extinction functions
}
\details{
Colonisation/extinction functions take a number of forms. Each form has its own
set of parameters that are provided when the function is created (e.g., when creating a species).
These functions represent the niche of a species, so the parameters are fixed.

Regardless of their form, the functions all have the same signature.

The parameter lists contain different named items depending on the type of function
\itemize{
\item 'constant'
\itemize{
\item \code{scale}: the height of the function
}
\item 'linear'
\itemize{
\item \code{a}: the y-intercept
\item \code{b}: the slope
}
\item 'gaussian'
\itemize{
\item \code{scale}: the maximum value of the curve
\item \code{mean}: the centre of the curve, x-value where y == \code{scale}
\item \code{sd}: the width of the curve
}
}
}
\examples{
fun = ce_linear(parm=list(a=0, b = 1))
R = seq(0,1,0.1)
fun(R)
}
