% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{ceprob}
\alias{ceprob}
\alias{col_prob}
\alias{ext_prob}
\title{Colonisation and extinction probability, given the current state of the system}
\usage{
col_prob(comm, network, dt, components = FALSE)

ext_prob(comm, network, dt, components = FALSE)
}
\arguments{
\item{comm}{A metacommnunity}

\item{network}{A river network}

\item{dt}{Time interval}

\item{components}{Boolean, normally FALSE, see 'details'}
}
\value{
Normally, a site by species matrix of colonisation/extinction probabilities, but see 'details'
}
\description{
Colonisation and extinction probability, given the current state of the system
}
\details{
For debugging and tuning, it can be useful to examine the individual components of the
colonisation/extinction rate by setting \code{components = TRUE}, in which case a named list of site
by species matrices is returned, one element per component.
}
