% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{check_par}
\alias{check_par}
\alias{.check_scale}
\alias{.check_breadth}
\alias{.check_breadth_multi}
\alias{.check_r_scale}
\title{Check parameter dimensions}
\usage{
.check_scale(p, nsp)

.check_breadth(p, nsp, nrx)

.check_breadth_multi(p, nsp, nrx)

.check_r_scale(p, nsp, nrx)
}
\description{
Check parameter dimensions
}
\keyword{internal}
