% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{competition}
\alias{competition}
\title{Produce a default competition matrix}
\usage{
competition(sp, xmin, xmax)
}
\arguments{
\item{sp}{A \code{\link[=species_pool]{species_pool()}}}

\item{xmin}{The minima for the resources}

\item{xmax}{The maxima for resources}
}
\value{
a matrix giving pairwise competition coefficients
}
\description{
Competition is defined by computing the overlap of species niches, which are in turn defined as c(R) - m(R).
This function computes the niche for each species, then integrates for every pairwise pair
}
\examples{
spp = flume:::species_pool()
competition(spp, 0, 1)

}
