% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{discharge}
\alias{discharge}
\alias{discharge.river_network}
\alias{discharge<-.river_network}
\alias{area.river_network}
\alias{area<-.river_network}
\title{Get and set discharge and cross-sectional area of a river network}
\usage{
\method{discharge}{river_network}(x, type = c("raw", "current", "history"))

\method{discharge}{river_network}(x) <- value

\method{area}{river_network}(x, type = c("raw", "current", "history"))

\method{area}{river_network}(x) <- value
}
\arguments{
\item{x}{A \code{\link[=river_network]{river_network()}}}

\item{type}{The type of record to get; either the raw data, the current state, or the whole
history}

\item{value}{Discharge/cross-sectional area (see 'details')}
}
\value{
A discharge vector

Cross-sectional area vector
}
\description{
Get and set discharge and cross-sectional area of a river network
}
\details{
There are three possible models for discharge/cross-sectional area, depending on the
kind of input provided.

If discharge is provided as a vector or a single-column matrix, then discharge will be treated
as constant throughout any simulations. In this case, then length of the input must equal the
number of nodes in the river network.

If discharge is provided as a matrix with ncol > 1, then a variable discharge model will be used.
The input in this case is a hydrograph, where the rows are nodes in the river network and the
columns are time steps. Time steps will be recycled, so if a model is run for a more steps than
is present in the discharge matrix, then discharge will restart at the first columns of the
matrix.

The final possibility is to provide a function that takes a single parameter, the time step, as
input and returns a discharge vector with one element per node in the river network.

Cross sectional area can either be missing or provided in the same format as discharge. If
missing, then cross sectional area will be computed using hydrological scaling relationships;
see \code{\link[=geometry]{geometry()}} for details.

Note that if cross-sectional area is provided, then the discharge model is changed (e.g., from
constant to variable), it is strongly recommended (and not checked!) that the user ALSO update
cross-sectional area to match the new model, otherwise undefined behavior will occur.

\code{area(x) <- NULL} will delete any provided cross sectional area and revert to the default
behavior, computing area using \code{\link[=geometry]{geometry()}}
}
