% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.r
\name{dispersal}
\alias{dispersal}
\alias{dispersal_custom}
\title{Generate scenarios for species' dispersal abilities}
\usage{
dispersal_custom(nsp, alpha = 0.05, beta = 0.5)
}
\arguments{
\item{nsp}{Number of species}

\item{alpha}{A scalar or a vector, active dispersal ability for each species}

\item{beta}{A scalar or a vector, passive dispersal ability for each species}

\item{...}{Additional named dispersal parameters to override the defaults in
\code{dispersal_custom()}.}
}
\value{
A list of dispersal parameters
}
\description{
Generate scenarios for species' dispersal abilities
}
\examples{
dispersal_custom(nsp = 3)
}
