% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{.check_state}
\alias{.check_state}
\title{Check state matrix for errors}
\usage{
.check_state(x, nn, nr = NA)
}
\arguments{
\item{x}{The state matrix to check}

\item{nn}{The number of nodes in the network}

\item{nr}{The number of resources, if NA, this is not checked}
}
\description{
Check state matrix for errors
}
\keyword{internal}
