% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{.compute_comp_matrix}
\alias{.compute_comp_matrix}
\title{Make a competition matrix}
\usage{
.compute_comp_matrix(sp, xmin, xmax)
}
\arguments{
\item{sp}{A \code{\link[=species_pool]{species_pool()}}}

\item{xmin}{The minimum for integration}

\item{xmax}{The maximum for integration}
}
\value{
a matrix giving pairwise competition coefficients
}
\description{
Make a competition matrix
}
\keyword{internal}
