% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{.default_plot_pool_options}
\alias{.default_plot_pool_options}
\title{Set default plot options when not user-specified}
\usage{
.default_plot_pool_options(...)
}
\description{
Set default plot options when not user-specified
}
\keyword{internal}
