% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{.dRdt_params}
\alias{.dRdt_params}
\title{Prepare parameter list for dRdt}
\usage{
.dRdt_params(C, N)
}
\arguments{
\item{C}{a metacommunity}

\item{N}{a river network}
}
\value{
Parameter list, to be passed to dRdt function
}
\description{
Prepare parameter list for dRdt
}
\keyword{internal}
