% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{.generate_ce_params}
\alias{.generate_ce_params}
\title{Create parameter sets for CE functions}
\usage{
.generate_ce_params(type, n, scale, xmin, xmax)
}
\description{
Create parameter sets for CE functions
}
\keyword{internal}
