% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{.get_hydro_attr}
\alias{.get_hydro_attr}
\title{Pull out discharge/cross sectional area}
\usage{
.get_hydro_attr(x, attr = c(".Q", ".area"), type)
}
\arguments{
\item{x}{A river network}

\item{attr}{Which attribute to get}
}
\description{
Pull out discharge/cross sectional area
}
\keyword{internal}
