% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{.pairwise_comp}
\alias{.pairwise_comp}
\title{Generate a function to integrate to determine competition between two species}
\usage{
.pairwise_comp(s1, s2)
}
\arguments{
\item{sp1}{First species}

\item{sp2}{Second species}
}
\description{
Generate a function to integrate to determine competition between two species
}
\keyword{internal}
