% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume.r
\name{.reshape_sim}
\alias{.reshape_sim}
\title{Get a long-format occupancy by time by reach by network dataset}
\usage{
.reshape_sim(x, variable = c("species", "resources"))
}
\description{
Get a long-format occupancy by time by reach by network dataset
}
\keyword{internal}
