% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{dRdt}
\alias{dRdt}
\title{Compute time-derivative for resources}
\usage{
dRdt(comm, network, components = FALSE)
}
\arguments{
\item{comm}{A metacommnunity}

\item{network}{A river network}

\item{components}{Boolean, normally FALSE, see 'details'}
}
\value{
Normally, a site by resource matrix of resource fluxes
}
\description{
Compute time-derivative for resources
}
\details{
For debugging and tuning, it can be useful to examine the individual components of the
derivative rate by setting \code{components = TRUE}, in which case a named list of site
by resource matrices is returned, with separate transport and reaction components.
}
