% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume-package.r
\docType{package}
\name{flume-package}
\alias{flume-package}
\title{flume: A Model For Fluvial Meta-Ecosystems}
\description{
Implements a model for fluvial meta-ecosystems.
}
\section{Key functions}{

\itemize{
\item \code{\link[=metacommunity]{metacommunity()}} Create species pools and associated attributes
\item \code{\link[=river_network]{river_network()}} Build a river network object
\item \code{\link[=flume]{flume()}} Creates a model
\item \code{\link[=run_simulation]{run_simulation()}} Runs the simulation
}
}

\author{
\strong{Maintainer}: Matthew Talluto \email{matthew.talluto@uibk.ac.at} (\href{https://orcid.org/0000-0001-5188-7332}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Selin Kubilay \email{D.S.Kubilay@tudelft.nl} [contributor]
  \item Thomas Fuß \email{thomas.fuss@uibk.ac.at} [data contributor]
}

}
\keyword{internal}
