% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume.r
\name{flume}
\alias{flume}
\title{Create a FLUvial Meta-Ecosystem model}
\usage{
flume(comm, network, dt)
}
\arguments{
\item{comm}{A \code{\link[=metacommunity]{metacommunity()}}}

\item{network}{A \code{\link[=river_network]{river_network()}}}

\item{dt}{The time step for the model, in the same time units as all fluxes}
}
\value{
An S3 object of class 'flume'
}
\description{
Create a FLUvial Meta-Ecosystem model
}
