% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume-package.r
\docType{data}
\name{flume_networks}
\alias{flume_networks}
\title{Flume river networks for 5 different rivers}
\format{
A \code{list}, with members:
\describe{
\item{kamp}{Kamp river, Austria}
\item{mara}{Mara river, Kenya}
\item{thur}{Thur river, Switzerland}
\item{vjosa}{Vjosa river, Greece & Albania}
\item{ybbs}{Ybbs river, Austria}
}
}
\usage{
flume_networks
}
\description{
Flume river networks for 5 different rivers
}
\keyword{datasets}
