% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume-package.r
\docType{package}
\name{flume_package}
\alias{flume_package}
\title{flume: A package for modelling FLUvial Meta-Ecosystems}
\description{
Implementation of a theoretical/mechanistic model for exploring fluvial meta-ecosystems.
}
\section{Key functions}{

\itemize{
\item \code{\link[=metacommunity]{metacommunity()}} Create species pools and associated attributes
\item \code{\link[=river_network]{river_network()}} Build a river network object
\item \code{\link[=flume]{flume()}} Creates a model
\item \code{\link[=run_simulation]{run_simulation()}} Runs the simulation
}
}

