% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume.r
\name{flumestats}
\alias{flumestats}
\alias{occupancy}
\alias{resource_summary}
\title{Compute post simulation statistics}
\usage{
occupancy(x, type = c("network", "reach"))

resource_summary(x)
}
\arguments{
\item{x}{A \code{\link[=flume]{flume()}}}
}
\value{
A \link[=data.table]{data.table::data.table()} with the desired summary statistics
}
\description{
Compute post simulation statistics
}
