% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.r
\name{generate_niches}
\alias{generate_niches}
\title{Create niches for a metacommunity}
\usage{
generate_niches(nsp, nrx, xmin = rep(0, nrx), xmax = rep(1, nrx))
}
\arguments{
\item{nsp}{The number of species}

\item{nrx}{The number of niche axes}

\item{xmin}{The minimum value(s) to use for generating niches}

\item{xmax}{The maximum value(s) for generating niches}

\item{k}{The number of niche dimensions}
}
\value{
A list with location, breadth, and scale parameters to pass to \code{\link[=metacommunity]{metacommunity()}}.
}
\description{
Create niches for a metacommunity
}
\details{
For a single niche axis, niches will be evenly spaced along the axis. For multiple
dimensions, niches locations are randomly generated
}
