% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{geometry}
\alias{geometry}
\title{Compute hydraulic geometry from discharge}
\usage{
geometry(Q)
}
\arguments{
\item{Q}{A vector of discharge values.}
}
\value{
A data frame with discharge, velocity, depth, and width
}
\description{
Compute hydraulic geometry from discharge
}
\references{
Raymond, PA et al. 2012. Scaling the gas transfer velocity and hydraulic
geometry in streams and small rivers. \emph{Limnology and Oceanography: Fluids and
Environments}. \strong{2}:41-53.
}
