% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche.r
\name{get_pars}
\alias{get_pars}
\alias{get_pars,NicheFun-method}
\alias{get_pars,ConstantNicheFun-method}
\alias{get_pars,GaussianNicheFun-method}
\title{Get niche Parameters}
\usage{
get_pars(object)

\S4method{get_pars}{NicheFun}(object)

\S4method{get_pars}{ConstantNicheFun}(object)

\S4method{get_pars}{GaussianNicheFun}(object)
}
\description{
Get niche Parameters
}
\keyword{internal}
