% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume-package.r
\docType{data}
\name{kamp}
\alias{kamp}
\title{Flume river network and associated data for the Kamp River catchment in Austria}
\format{
A \code{list}, with members:
\describe{
\item{network}{A \link{river_network}}
}
}
\usage{
kamp
}
\description{
Flume river network and associated data for the Kamp River catchment in Austria
}
\keyword{datasets}
