% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{lateral_discharge}
\alias{lateral_discharge}
\title{Compute lateral discharge}
\usage{
lateral_discharge(x)
}
\arguments{
\item{x}{A river network}

\item{tm}{A time step}
}
\value{
A vector of discharge values
}
\description{
Assumed to be the difference between Q of a site and the sum of Q of upstream sites
}
