% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{length.river_network}
\alias{length.river_network}
\title{Number of nodes in a river network}
\usage{
\method{length}{river_network}(x)
}
\arguments{
\item{x}{A river network}
}
\value{
The number of nodes in the network
}
\description{
Number of nodes in a river network
}
