% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{niche}
\alias{niche}
\alias{f_niche.species}
\alias{f_niche.metacommunity}
\title{Functions for determining the niche of species}
\usage{
\method{f_niche}{species}(x, R)

\method{f_niche}{metacommunity}(x, R)
}
\arguments{
\item{x}{A \code{\link[=species]{species()}} or \code{\link[=metacommunity]{metacommunity()}}}

\item{R}{A site by resource matrix}
}
\value{
A vector (for a single species) or matrix (for a metacommunity) giving the value(s)
of the niche at each site
}
\description{
Functions for determining the niche of species
}
\examples{
comm = metacommunity()
st = matrix(seq(0, 1, length.out = 4), ncol = 1, dimnames = list(NULL, 'R'))

# for a single species
f_niche(comm$species[[1]], st)

# for multiple species
f_niche(comm, st)
}
