% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.r
\name{niches}
\alias{niches}
\alias{niches_custom}
\alias{niches_uniform}
\alias{niches_random}
\title{Generate scenarios for species niches}
\usage{
niches_custom(
  nsp,
  nr,
  nn = nr,
  location,
  breadth = 1,
  scale_c = 0.5,
  scale_e = 0.2,
  dry_c = 0,
  dry_e = Inf,
  r_use = 0.05,
  static = integer(),
  alpha = 0.05,
  beta = 0.5,
  r_trans = identity
)

niches_uniform(nsp = 2, nr = 1, r_lim = c(0, 1), ...)

niches_random(nsp = 2, nr = 1, r_lim = c(0, 1), ...)
}
\arguments{
\item{nsp}{Number of species}

\item{nr}{Number of resources}

\item{nn}{Number of niche dimensions}

\item{location}{Location parameter for the niche, see 'details'.}

\item{breadth}{Niche breadth, see 'details'.}

\item{scale_c}{Niche height for colonisation, see 'details'.}

\item{scale_e}{Niche height for extinction, see 'details'.}

\item{dry_c}{Factor to multiply colonisation rate by when a reach is dry}

\item{dry_e}{Factor to multiply extinction rate by when a reach is dry}

\item{r_use}{Resource use scaling parameter, see 'details}

\item{static}{Vector of resource indices that are static, default is empty vector}

\item{alpha}{A scalar or a vector, active dispersal ability for each species}

\item{beta}{A scalar or a vector, passive dispersal ability for each species}

\item{r_trans}{A function for transforming niche dimensions, default is identity}

\item{...}{Additional named niche parameters to override the defaults in \code{niches_custom()}.}
}
\value{
A list of niche parameters, suitable for passing on to \code{\link[=species]{species()}}.
}
\description{
Generate scenarios for species niches
}
\details{
Scenarios handle the complexity of generating niches for you. For custom niches, there is some
in allowable parameter values and dimensions. Details follow.

\code{location}: Required for custom niches; the location of the niche optimum for each
species-resource combination.
Must be a matrix with one row per species (\code{nsp}), one column per niche axis (\code{nr}). For a
single niche axis, a vector with one entry per species is also accepted. The location is the
mean of the Gaussian colonisation function.

\code{breadth}: Niche breadth, or the standard deviation of the Gaussian colonisation function;
larger values indicate species can occur in a wider variety of environments. For a single
niche dimension, either a scalar (all species have the same breadth) or a vector of length \code{nsp}.
For multivariate niches, the following are possible:
\itemize{
\item A \strong{scalar}: all species have the same breadth in each niche dimension
\item A \strong{vector} of length \code{nsp}: each species has a different breadth, but breadth is the same
for each niche axis.
\item A \strong{vector} of length \code{nr}: niche breadth varies by axis, but all species have the same
breadth for each niche axis. If \code{nr == nsp}, then this option is not possible, instead the
former applies.
\item A \strong{matrix} of \code{nsp} rows and \code{nr} columns: each species-resource combo has a unique
breadth, but all resources are orthogonal. In other words, performance for a given resource
cannot depend on the concentration of any other resource.
\item A \strong{list} of length \code{nsp}, each element is a square symmetric \code{nr} by \code{nr} matrix. Similar
to the matrix above, but a full variance-covariance matrix is supplied for each species,
describing the breadth of the niche along each axis but also how the axes covary.
}

For ratio niche dimensions, note that \code{nr} will be larger than the number of niche dimensions
specified in \code{location} and \code{breadth}.  Ratio dimensions will always be inserted at the end
of the resource matrix, regardless of the position of the original resources, and it is necessary
that the \code{location} and \code{breadth} parameters reflect this. For example, if the niche dimensions
for three resources are the r1:r2 ratio and r3, the \code{location} parameter should be of length 2,
with the first entry being the location for r3, and the second for r1:r2. To avoid confusion,
it is recommended (but not required) that users only construct ratio niches from resources that
are specified at the end.

\code{scale_c}, \code{scale_e}: The (relative) height of the Gaussian colonisation function or
the constant extinction function, must be a positive real number.
Can be supplied as a scalar (all species have the same scale) or a vector of length \code{nsp}.
If missing a default value of 0.5 (for colonisation) or 0.2 (for extinction) will be used.

\code{r_use}: The scale of resource consumption; larger values indicate
faster consumption of resources, see \code{\link[=ruf]{ruf()}} for more details. This can be a scalar (all species
consume all resources at the same rate relative to niche position), a vector of length \code{nsp}
(species behave differently, but all resources within species are consumed identically),
a vector of length \code{nr} (all species have identical behaviour, but each resource is consumed at
a different rate), or a matrix with \code{nsp} rows and \code{nr} columns. Note that static resources
ignore this parameter.
}
\examples{
niches_uniform(nsp = 4)
niches_custom(nsp = 2, nr = 2, location = matrix(c(1,2,3,4), nrow=2))
}
