% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{nparams}
\alias{nparams}
\alias{niche_par.species}
\alias{niche_par.metacommunity}
\title{Functions for accessing species niche parameters}
\usage{
\method{niche_par}{species}(
  x,
  par = c("location", "breadth", "sd", "scale", "r_use", "niche_max", "dry")
)

\method{niche_par}{metacommunity}(
  x,
  par = c("location", "breadth", "sd", "scale", "r_use", "niche_max", "dry")
)
}
\arguments{
\item{x}{A species or metacommunity}

\item{par}{The parameter to retrieve}
}
\description{
Functions for accessing species niche parameters
}
