% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{plot.species}
\alias{plot.species}
\title{Plot species niches}
\usage{
\method{plot}{species}(x, R, axis = 1, ...)
}
\arguments{
\item{x}{A species}

\item{R}{a resource state matrix}

\item{axis}{Which resource axis (i.e., column in R) to plot along}
}
\description{
Plot species niches
}
