% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{prevalence}
\alias{prevalence}
\title{Compute current prevalence for all species in a river network}
\usage{
prevalence(x)
}
\arguments{
\item{x}{A river network}
}
\value{
A site by species matrix; values indicate the prevalence in surrounding sites
}
\description{
Compute current prevalence for all species in a river network
}
