% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.r
\name{r_transform}
\alias{r_transform}
\alias{ratio_transform}
\title{Transform resources into niche dimensions}
\usage{
ratio_transform(ratio)
}
\description{
Transform resources into niche dimensions
}
\keyword{internal}
