% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{random_community}
\alias{random_community}
\title{Generate a random community}
\usage{
random_community(rn, mc, prevalence = 0.25)
}
\arguments{
\item{rn}{A \code{\link[=river_network]{river_network()}}}

\item{mc}{A \code{\link[=metacommunity]{metacommunity()}}}

\item{prevalence}{A vector of length 1 or \code{length(mc)}, what proportion of sites should be occupied on average by each species.}
}
\value{
A site by species matrix, with sites taken from \code{rc} and species from \code{mc}
}
\description{
Generate a random community
}
\examples{
Q = rep(1, 4)
adj = matrix(0, nrow = 4, ncol = 4)
adj[1,2] = adj[2,3] = adj[4,3] = 1
rn = river_network(adj, Q)
comm = community()
site_by_species(rn) = random_community(rn, mc)
plot(rn, variable = "site_by_species")
}
