% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{reach_length}
\alias{reach_length}
\title{Length of every reach in the river network}
\usage{
reach_length(x)
}
\arguments{
\item{x}{A river network}
}
\value{
a vector of reach lengths
}
\description{
Currently, all reaches are assumed to have the same length
}
