% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{river_network}
\alias{river_network}
\title{Create a river network}
\usage{
river_network(adjacency, discharge, skip_checks = FALSE)
}
\arguments{
\item{adjacency}{An adjacency matrix; see 'details'}

\item{discharge}{A vector for discharge values, one per row/column in adjacency}

\item{skip_checks}{Logical; if true, no checks for valid topology will be performed}
}
\description{
Create a river network
}
\details{
The adjacency matrix must be a square matrix, with a row/column for each river network reach. Zero entries
indicate no adjacency between reaches. Nonzero entries indicate adjacency, with the value equal to the distance between
reaches.
}
