% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume.r
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation for a specified number of time steps}
\usage{
run_simulation(x, nt, reps, parallel = TRUE)
}
\arguments{
\item{x}{A \code{\link[=flume]{flume()}} object}

\item{nt}{The number of time steps}

\item{reps}{The number of replicate simulations to run; by default a single sim is run
on a new flume; for continuing simulations, the same number of replicates will be used}
}
\value{
A modified copy of \code{x}, with state updated with the results of the simulation.
}
\description{
This function begins with the current state of \code{x}, runs the model for \code{nt} steps,
then returns a copy of \code{x} with the updated state and state history.
}
