% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{species}
\alias{species}
\title{Creates species}
\usage{
species(c_type, e_type, c_par, e_par, alpha = 0, beta = 0)
}
\arguments{
\item{c_type}{Form of \link[=ce_funs]{colonisation function}}

\item{e_type}{Form of \link[=ce_funs]{extinction function}}

\item{c_par}{Named parameter list; constant parameters to include in the c/e functions, see \link[=ce_funs]{colonisation functions}}

\item{e_par}{Named parameter list; constant parameters to include in the c/e functions, see \link[=ce_funs]{extinction functions}}

\item{alpha}{Active dispersal ability}

\item{beta}{Passive dispersal ability}
}
\value{
An S3 object of class 'species', which contains the following named elements:
\code{col}: The colonisation function, takes a state matrix R and returns a vector of colonisation rates
\code{ext}: The extinction function
\code{c_par}: colonisation niche parameters
\code{e_par}: extinction niche parameters
\code{alpha}: Active dispersal ability
\code{beta}: Passive dispersal ability
}
\description{
Creates species
}
\details{
Creates a species, which is defined by its niche parameters and its colonisation/extinction functions

Note that for gaussian functions in multiple variables, \code{scale} is always a scalar, but
\code{mean} must be a location vector of length \code{nx}, and \code{sd} must be a variance covariance matrix
with \code{dim=c(nx,nx)}.

The resource use function is set automatically by default, but can be set to an arbitrary function; see
\link[=ruf]{resource use functions} for details on how this function should behave.
}
\examples{
sp = flume:::species('linear', 'constant', list(a=0, b=1), e_par = list(scale = 0.2))
plot(sp)
}
