% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{species_pool}
\alias{species_pool}
\title{Generates a species pool with all possible species in a metacommunity}
\usage{
species_pool(n_species = 2, nx = 1, c_type = "linear",
  e_type = "constant", scale = c(1, 0.2), xmin = rep(0, nx),
  xmax = rep(1, nx))
}
\arguments{
\item{n_species}{The number of species}

\item{nx}{The number of niche axes}

\item{c_type}{The type of \link[=ce_funs]{colonisation function}}

\item{e_type}{The type of \link[=ce_funs]{extinction function}}

\item{scale}{A vector, length 2, with the scales (maximum values) for the colonisation and extinction functions}

\item{xmin}{Vector, length=nx, the minimum values for each resource for scaling the fitness functions}

\item{xmax}{Vector, length=nx, the maximum values for each resource for scaling the fitness functions}
}
\value{
A list of \code{\link[=species]{species()}}
}
\description{
In general species in a species pool all have the same niche shapes (c_type and e_type). Species with
different shapes in possible in principle, however; in this case a species pool can be created with
a set of default species and then individual species added or replaced using \code{\link[=species]{species()}}.
}
\examples{
## create a default 2-species pool
spp = flume:::species_pool()
}
