% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.r
\name{summarise}
\alias{summarise}
\title{Methods for summarising simulation results}
\usage{
summarise(
  x,
  by = c("time", "reach"),
  stat,
  quantile = c(0.05, 0.95),
  t_steps = NA
)
}
\arguments{
\item{x}{A fitted flume model}

\item{by}{The margin(s) across which to summarise; see 'details'}

\item{stat}{The statistic(s) to compute; see 'details'}

\item{quantile}{The quantiles to compute; see 'details'}

\item{t_steps}{What time steps to include in the summary; default \code{NA} will include all}
}
\value{
A data.table with the desired summary statistics
}
\description{
Methods for summarising simulation results
}
\details{
The \code{by} parameter controls how detailed of a summary to create; adding more
margins creates a more detailed summary. The default behaviour provides statistics for each
time step and each reach, with quantiles computed across replicate networks. Allowable values:
\itemize{
\item time
\item reach
\item species (incompatible with resources; not yet implemented)
\item resources (incompatible with species; not yet implemented)
}

The statistics desired interact with the margins selected; certain values for \code{stat} are only
compatible with specific values within \code{by}. If missing, a default set will be selected given the
value in \code{by}. Allowable values, with compatible margins in parentheses, and required margins
in bold:
\itemize{
\item occupancy (\strong{species}, reach, time)
\item EF (resources, reach, time)
\item richness (reach, time)
\item concentration (\strong{resources}, reach, time)
}

If the fitted model includes multiple replicate networks, then summarise by default returns
upper and lower quantiles along with the median, summarised across networks.
Only two quantiles may be provided. If quantile is NULL, then all replicates will be returned.
}
