% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume-package.r
\docType{data}
\name{algae}
\alias{algae}
\title{Algal metacommunity N:P example dataset}
\format{
A \code{\link[=flume]{flume()}}
\describe{
\item{species}{Five algal ASVs}
\item{sites}{Sampled locations in the Bodingbach catchment}
}
}
\usage{
algae
}
\description{
Algal metacommunity N:P example dataset
}
\keyword{datasets}
