% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume-package.r
\docType{data}
\name{algae}
\alias{algae}
\title{Algal metacommunity N:P example dataset}
\format{
A list
\describe{
\item{niches}{A table giving the niche location and breadth (in terms of N:P) for five species}
}
}
\usage{
algae
}
\description{
Algal metacommunity N:P example dataset
}
\keyword{datasets}
