% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume-package.r
\docType{data}
\name{algae}
\alias{algae}
\title{Algal metacommunity N:P dataset}
\format{
A list
\describe{
\item{niche_params}{A data.frame giving the niche location, breadth, and scale (in terms of N:P) for 40 species}
\item{r0}{Starting resource state for N and P}
\item{sp0}{Starting locations for all species}
}
}
\usage{
algae
}
\description{
Algal metacommunity N:P dataset
}
\keyword{datasets}
