% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_funs.r
\name{ce_funs}
\alias{ce_funs}
\alias{ce_linear}
\alias{ce_constant}
\alias{ce_gaussian}
\title{Colonisation and extinction functions}
\usage{
ce_linear(parm)

ce_constant(scale)

ce_gaussian(location, breadth, scale)
}
\arguments{
\item{scale}{Height of the col/ext function}

\item{location}{Location optimum for functions with an optimum}

\item{breadth}{Breadth of the function (e.g., standard deviation or vcv matrix for gaussian)}

\item{x}{state variable matrix; input for c/e functions; one row per site, one column per variable.}
}
\value{
For c/e functions, a vector of colonisation/extinction rates with length == \code{nrow(x)}
For all others, a c/e function of the desired form
}
\description{
Colonisation and extinction functions
}
\details{
Colonisation/extinction functions take a number of forms. Each form has its own
set of parameters that are provided when the function is created (e.g., when creating a species).
These functions represent the niche of a species, so the parameters are fixed.
}
\examples{
fun = ce_linear(parm=list(a=0, b = 1))
R = seq(0,1,0.1)
fun(R)
}
