% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check_par}
\alias{check_par}
\alias{.check_scale}
\alias{.check_ratio}
\alias{.check_breadth}
\alias{.check_breadth_multi}
\alias{.check_r_use}
\alias{.check_species_params}
\title{Check parameter dimensions}
\usage{
.check_scale(p, nsp)

.check_ratio(r, nr)

.check_breadth(p, nsp, nrx)

.check_breadth_multi(p, nsp, nrx)

.check_r_use(p, nsp, nrx, static)

.check_species_params(x)
}
\arguments{
\item{p}{The parameter to check}

\item{nsp}{The number of species}

\item{nrx}{The number of resource/niche dimensions}
}
\description{
Check parameter dimensions
}
\keyword{internal}
