% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{.compute_comp_matrix}
\alias{.compute_comp_matrix}
\title{Make a competition matrix}
\usage{
.compute_comp_matrix(x, comp_scale)
}
\arguments{
\item{x}{A \code{\link[=metacommunity]{metacommunity()}}}

\item{comp_scale}{The competition scale}
}
\value{
a matrix giving pairwise competition coefficients
}
\description{
Make a competition matrix
}
\keyword{internal}
