% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume.r
\name{.do_sim}
\alias{.do_sim}
\title{Worker function for running simulations in parallel}
\usage{
.do_sim(network, comm, dt, nt)
}
\arguments{
\item{network}{A river network}

\item{comm}{A metacommunity}

\item{dt}{The size of the time step; one "generation" for the metacommunity}

\item{nt}{The number of time steps}
}
\value{
A modified copy of \code{network} with the results of the simulation
}
\description{
Worker function for running simulations in parallel
}
\keyword{internal}
