% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{.generate_ce_params}
\alias{.generate_ce_params}
\title{Create parameter sets for CE functions}
\usage{
.generate_ce_params(type, n, scale, xmin, xmax, k = length(xmin))
}
\arguments{
\item{type}{The type of c-e function}

\item{n}{The number of species}

\item{scale}{The maximum y-value of the function}

\item{xmin}{The minimum value(s) to use for generating niches}

\item{xmax}{The maximum value(s) for generating niches}

\item{k}{The number of niche dimensions}
}
\value{
A parameter set for the c-e function chosen
}
\description{
Create parameter sets for CE functions
}
\keyword{internal}
