% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.r
\name{.output_table}
\alias{.output_table}
\title{Extract a variable across an entire network or series of networks}
\usage{
.output_table(x, v, t_steps)
}
\arguments{
\item{x}{A \code{\link[=river_network]{river_network()}} or \code{\link[=flume]{flume()}}}

\item{v}{The variable to extract}

\item{t_steps}{What time steps to include in the summary; default \code{NA} will include all}
}
\value{
A data.table
}
\description{
Extract a variable across an entire network or series of networks
}
\keyword{internal}
