% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{.validate_adjacency}
\alias{.validate_adjacency}
\title{Perform checks on a river network adjacency matrix}
\usage{
.validate_adjacency(adjacency)
}
\arguments{
\item{adjacency}{An adjacency matrix}
}
\value{
Logical, TRUE if the validation passes
}
\description{
Perform checks on a river network adjacency matrix
}
\keyword{internal}
