% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{dRdt}
\alias{dRdt}
\title{Compute time-derivative for resources}
\usage{
dRdt(t, R, params, components = FALSE)
}
\arguments{
\item{t}{Time variable}

\item{R}{A resource state matrix; stored as a vector in column major order}

\item{params}{Named list; additional parameters, see 'details'}
}
\value{
Normally, a site by resource matrix of resource fluxes, in units of
mass * volume^-1 * time^-1
}
\description{
Compute time-derivative for resources
}
\details{
The required params for the resource derivative are as follows:
\itemize{
\item S: a site by species matrix
\item Q: a site by discharge vector
\item A: a site by cross-sectional area vector
\item l: a site by length vector
\item lQ: site by lateral input discharge
\item lR: site by lateral resource concentration
\item i_static: indices of static resources
\item adj: the network adjacency matrix
\item comm: a \code{\link[=metacommunity]{metacommunity()}}
\item components: logical, optional, default FALSE
}

If \code{params$components = TRUE}, instead of computing the derivative the function returns
the individual components, including the use by species, downstream transport, and input from
upstream. The units in this case will be mass/time
}
