% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume.r
\docType{package}
\name{flume}
\alias{flume}
\alias{flume-package}
\title{flume: A package for modelling FLUvial Meta-Ecosystems}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Key functions}{
NA
}

