% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{metacommunity}
\alias{metacommunity}
\title{Greate a community for simulations}
\usage{
metacommunity(n_species = 2, nx = 1, xmin = rep(0, nx), xmax = rep(1, nx), ...)
}
\arguments{
\item{n_species}{The number of species}

\item{nx}{The number of niche axes}

\item{xmin}{Vector, length=nx, the minimum values for each resource for scaling the fitness functions}

\item{xmax}{Vector, length=nx, the maximum values for each resource for scaling the fitness functions}

\item{...}{Additional parameters to pass to \code{\link[=species_pool]{species_pool()}}}
}
\value{
An object of class \code{community}, with the following items:
\itemize{
\item \code{species} a \link[=species_pool]{species pool}
\item \code{competition} a \code{\link[=competition]{competition()}} matrix
}

Additionally, the following attributes:
\itemize{
\item \code{xmin} the minimum value of the environmental gradient
\item \code{xmax} the maximum value of the environmental gradient
}
}
\description{
Greate a community for simulations
}
\examples{
comm = community()
plot(comm)
}
