% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{metacommunity}
\alias{metacommunity}
\title{Create a metacommunity}
\usage{
metacommunity(
  nsp = 2,
  nr = 1,
  niches = niches_uniform,
  dispersal = dispersal_custom,
  sp_names = paste0("sp", 1:nsp),
  r_names = paste0("r", 1:nr),
  niche_args = list(),
  dispersal_args = list()
)
}
\arguments{
\item{niches}{The \href{niches}{niche scenario} to use.}

\item{dispersal}{The \href{dispersal}{dispersal scenario} to use}

\item{sp_names}{A vector of species names}

\item{r_names}{A vector of resource names}

\item{niche_args}{A list of named arguments to be passed to the niche function}

\item{dispersal_args}{A list of named arguments to be passed to the dispersal function}
}
\value{
A metacommunity object
}
\description{
Create a metacommunity
}
\details{
The metacommunity describes the possible biological space for a \code{flume} model. It
consists of a few named elements:
\itemize{
\item \code{species}: a list of \code{\link[=species]{species()}} objects, describing the niche and dispersal parameters
for each possible species
\item \code{competition}: A square species interaction matrix describing the effect of each species
on each other species; generated using the overlap between species'
\href{f_niche.species()}{fundamental niches}.
\item \code{boundary}: A function that returns a site by species matrix giving the colonisation flux
from outside the river network; by default returns zero for all sites/species.
}

Each species' fundamental niche is defined as the difference between a Gaussian colonisation
function and a constant extinction function; a species will be present when c - e > 0.
Niches can be automatically generated (see \code{\link[=generate_niches]{generate_niches()}}; alternatively, to specify
the parameters manually, the following can be specified.
}
\examples{
\donttest{
	# defaults
	mc = metacommunity()

	# multiple species
	mc = metacommunity(location = c(1, 2))

	# multiple resources
	nlim = matrix(c(0, 0, 1, 1), ncol = 2) # we need to specify the limits manually
	mc = metacommunity(location = matrix(c(1, 2), ncol = 2), niche_lim = nlim)

	# multiple species and resources with variance covariance matrices
	loc = matrix(c(1, 2, 1, 3, 2, 4), ncol = 2) # 3 species, 2 resources
	br_sp1 = matrix(c(1, 0.2, 0.2, 1), ncol = 2) # vcv matrix for one species
	breadth = list(br_sp1, br_sp1, br_sp1) # assume the same matrix for all species
	mc = metacommunity(location = loc, breadth = bre, niche_lim = nlim)
}
}
