% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.r
\name{niches}
\alias{niches}
\alias{niches_custom}
\alias{niches_uniform}
\alias{niches_random}
\title{Generate scenarios for species niches}
\usage{
niches_custom(
  nsp,
  nr,
  cfun = "gaussian",
  efun = c("constant", "inverse_gaussian"),
  location,
  breadth = 1,
  scale = list(col = 5e-06, ext = 2e-07),
  r_use = 5e-04,
  r_lim = matrix(rep(c(0, 1), each = nr), ncol = 2),
  static,
  ratio
)

niches_uniform(nsp = 2, nr = 1, r_lim = c(0, 1), ...)

niches_random(nsp = 2, nr = 1, r_lim = c(0, 1), ratio, ...)
}
\arguments{
\item{nsp}{Number of species}

\item{nr}{Number of resources}

\item{cfun, }{efun Niche function to use, see 'details'.}

\item{location}{Location parameter for the niche, see 'details'.}

\item{breadth}{Niche breadth, see 'details'.}

\item{scale}{Niche scale, see 'details'.}

\item{r_use}{Resource use scaling parameter, see 'details}

\item{r_lim}{Minimum and maximum possible values for each resource dimension; can be a vector
in the case of a single resource, otherwise a matrix with 2 columns and nr rows.}

\item{static}{Vector, the indices of resources that are static, \code{NA} if none. Any resources
that are not static are by definition dynamic}

\item{ratio}{A 2-column matrix; each row is the indices of a pair of resources. The ratio of
the first to the second will define a single niche dimension. If missing, no ratios used.}

\item{...}{Additional named niche parameters to override the defaults in \code{niches_custom()}.}
}
\value{
A list of niche parameters, suitable for passing on to \code{\link[=species]{species()}}.
}
\description{
Generate scenarios for species niches
}
\details{
Scenarios handle the complexity of generating niches for you. For custom niches, there is some
flexibility in allowable parameter values and dimensions.
}
\section{Niche parameters}{


\code{cfun}, \code{efun}: The niche function to use; for colonisation only gaussian is supported,
for extinction may be be constant (the default) or inverse_gaussian.

\code{location}: Required, no default if using \code{niches_custom}. The location of the niche optimum
for each species-resource combination. Must be a matrix with one row per species (\code{nsp}),
one column per niche axis (\code{nr}). For a single niche axis, a vector with one entry per
species is also accepted.

\code{breadth}: Niche breadth, or the standard deviation of the gaussian col/ext function;
larger values indicate species can occur in a wider variety of environments. For a single
niche dimension, either a scalar (all species have the same breadth) or a vector of length \code{nsp}.
For multivariate niches, the following are possible:
\itemize{
\item A \strong{scalar}: all species have the same breadth in each niche dimension
\item A \strong{vector} of length \code{nsp}: each species has a different breadth, but breadth is the same
for each niche axis.
\item A \strong{vector} of length \code{nr}: niche breadth varies by axis, but all species have the same
breadth for each niche axis. If \code{nr == nsp}, then this option is not possible, instead the
former applies.
\item A \strong{matrix} of \code{nsp} rows and \code{nr} columns: each species-resource combo has a unique
breadth, but all resources are orthogonal. In other words, performance for a given resource
cannot depend on the concentration of any other resource.
\item A \strong{list} of length \code{nsp}, each element is a square symmetric \code{nr} by \code{nr} matrix. Similar
to the matrix above, but a full variance-covariance matrix is supplied for each species,
describing the breadth of the niche along each axis but also how the axes covary.
\item A \strong{list} with two elements named \code{col} and \code{ext}, each element is a valid entry above,
in order to provide separate breadths for gaussian colonisation and extinction niches.
}

For ratio niche dimensions, note that \code{nr} will be larger than the number of niche dimensions
specified in \code{location} and \code{breadth}.  Ratio dimensions will always be inserted at the end
of the resource matrix, regardless of the position of the original resources, and it is necessary
that the \code{location} and \code{breadth} parameters reflect this. For example, if the niche dimensions
for three resources are the r1:r2 ratio and r3, the \code{location} parameter should be of length 2,
with the first entry being the location for r3, and the second for r1:r2. To avoid confusion,
it is recommended (but not required) that users only construct ratio niches from resources that
are specified at the end. Default value 1. If a gaussian curve is desired for both
colonisation and extinction, it

\code{scale}: Named list with two elements (\code{col} and \code{ext}) giving the (relative) heights of the colonisation
and extinction functions. Must be positive real numbers. Each element can be supplied
as a scalar (all species have the same scale) or a vector of length \code{nsp}.
If missing a default value of 5e-6 (for colonisation) or 2e-7 (for extinction) will be used.
For gaussian extinction niches, the scale is the height of the curve \emph{minimum}

\code{r_use}: The scale of resource consumption; larger values indicate
faster consumption of resources, see \code{\link[=ruf]{ruf()}} for more details. This can be a scalar (all species
consume all resources at the same rate relative to niche position), a vector of length \code{nsp}
(species behave differently, but all resources within species are consumed identically),
a vector of length \code{nr} (all species have identical behaviour, but each resource is consumed at
a different rate), or a matrix with \code{nsp} rows and \code{nr} columns. Note that static resources
ignore this parameter. Default value 5e-4.
}

\examples{
niches_uniform(nsp = 4)
niches_custom(nsp = 2, nr = 2, location = matrix(c(1,2,3,4), nrow=2))
}
