% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{plot.flume}
\alias{plot.flume}
\title{Plot the results of a flume simulation}
\usage{
\method{plot}{flume}(x, type = "occupancy")
}
\arguments{
\item{x}{A \code{\link[=flume]{flume()}}}

\item{type}{Which plot to produce; see 'details'.}
}
\value{
A ggplot2 object
}
\description{
Plot the results of a flume simulation
}
\details{
\code{flume} comes with a collection of default plots. All of them are built around the
\code{\link[=summarise]{summarise()}} function, which can be used to extract the data from a fitted \code{flume}. The
following plot \code{type}s are recognized (partial matches are allowed):\preformatted{* **occupancy**: (plots the proportion of sites occupied by each in the network against time,  
with quantiles if replciates are present)
}

All plots return a \code{ggplot}, so further customisation is easy.
}
