% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{plot.flume}
\alias{plot.flume}
\title{Plot the results of a flume simulation}
\usage{
\method{plot}{flume}(x, variable = c("occupancy", "resources"), type = c("network", "reach"))
}
\arguments{
\item{x}{A \code{\link[=flume]{flume()}}}

\item{variable}{Variable to plot; 'occupancy' plots occupancy by species over time, 'resources' plots
resource concentration over time}

\item{type}{How to compute the plot; either shows the average for the entire network, or plots by reach.}
}
\value{
A ggplot2 object
}
\description{
Plot the results of a flume simulation
}
