% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.r
\name{plot.metacommunity}
\alias{plot.metacommunity}
\title{Plot species independent stable envelopes}
\usage{
\method{plot}{metacommunity}(x, R, axis = 1, res = 100, ...)
}
\arguments{
\item{x}{A \code{\link[=metacommunity]{metacommunity()}}}

\item{R}{an optional resource state matrix}

\item{axis}{Which resource axis (i.e., column in R) to plot along}

\item{res}{Plotting resolution, how many points along the x-axis to plot; more produces smoother
lines.}
}
\description{
Plot species independent stable envelopes
}
