% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{plot.metacommunity}
\alias{plot.metacommunity}
\title{Plot species independent stable envelopes}
\usage{
\method{plot}{metacommunity}(x, axis = 1, res = 100, default_r, lwd = 1, xlim, ylim)
}
\arguments{
\item{x}{A \code{\link[=metacommunity]{metacommunity()}}}

\item{axis}{Which niche axis to plot along}

\item{res}{Plotting resolution, how many points along the x-axis to plot; more produces smoother
lines.}

\item{default_r}{Default values for niche dimensions not being plotted along. If not specified
the default is to use the midpoint of possible values for each niche axis.}

\item{xlim}{A vector of two, optional axis limits for plotting}

\item{ylim}{A vector of two, optional axis limits for plotting}
}
\description{
Plot species independent stable envelopes
}
