% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{plot.river_network}
\alias{plot.river_network}
\title{Plot a river network}
\usage{
\method{plot}{river_network}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=river_network]{river_network()}}}

\item{...}{Additional arguments to \code{\link[igraph:plot.igraph]{igraph::plot.igraph()}}}
}
\description{
Plot a river network
}
