% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{plot.species}
\alias{plot.species}
\title{Plot species niches}
\usage{
\method{plot}{species}(x, R, axis = 1, res = 100, lwd = 1)
}
\arguments{
\item{x}{A species}

\item{axis}{Which niche axis (i.e., column in R) to plot along}

\item{res}{Plotting resolution, how many points along the x-axis to plot; more produces smoother
lines.}
}
\value{
A ggplot2 object; additional modifications to the plot can be made using usual ggplot2
syntax.
}
\description{
Plot species niches
}
