% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.R
\name{river_network}
\alias{river_network}
\title{Create a river network}
\usage{
river_network(
  adjacency,
  discharge,
  area,
  length = 1,
  layout,
  site_names = rownames(adjacency),
  skip_checks = FALSE
)
}
\arguments{
\item{adjacency}{An adjacency matrix; see 'details'}

\item{discharge}{Optional discharge values, see \code{\link[=discharge.river_network]{discharge.river_network()}}}

\item{area}{Optional cross sectional area values, see \code{\link[=discharge.river_network]{discharge.river_network()}}}

\item{length}{The length of each reach, must be a single value (the model assumes all reaches
are the same length)}

\item{layout}{Optional, matrix of x-y coordinates used for plotting the network}

\item{site_names}{Optional vector of site names}

\item{skip_checks}{Logical; if true, no checks for valid topology will be performed}
}
\value{
An S3 object of class 'river_network', with the following members:
\itemize{
\item \code{adjacency} The adjacency matrix
\item \code{discharge} A discharge vector
\item \code{area} Cross sectional area at each node
\item \code{.state} The state history of the network; access with \link[=state.river_network]{state}.
}

Additionally, the \code{layout} attribute, if used, is a matrix of coordinates for plotting
}
\description{
Create a river network
}
\details{
The adjacency matrix must be a square matrix, with a row/column for each river network
reach. Zero entries indicate no adjacency between reaches. Nonzero entries indicate adjacency,
with the row indicating the upstream reach and the column the downstream
(i.e., adjacency\link{i,j} == 1 indicates water flows from i to j). Nonzero elements will be set
equal to one.
}
\examples{
Q = rep(1, 4)
adj = matrix(0, nrow = 4, ncol = 4)
adj[1,2] = adj[2,3] = adj[4,3] = 1
rn = river_network(adj, Q)
plot(rn)
}
