% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flume.r
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation for a specified number of time steps}
\usage{
run_simulation(x, nt, reps, parallel = TRUE, cores = parallel::detectCores())
}
\arguments{
\item{x}{A \code{\link[=flume]{flume()}} object}

\item{nt}{The number of time steps, see 'details'}

\item{reps}{The number of replicate simulations to run; by default a single sim is run
on a new flume; for continuing simulations, the same number of replicates will be used}
}
\value{
A modified copy of \code{x}, with state updated with the results of the simulation.
}
\description{
This function begins with the current state of \code{x}, runs the model for \code{nt} steps,
then returns a copy of \code{x} with the updated state and state history.
}
\details{
The time scale of flume is determiend by the combination of (a) the units of the
discharge variable (assumed generally to be m^3/s) and (b) the size of the time step (which is
specified with the \code{dt} parameter of the \code{\link[=flume]{flume()}} function).

At each time step of length \code{dt}, the model will simulate colonisations and extinctions assuming
that resource state is constant throughout the time step, and simulate resource dynamics
continuously via numerical integration. Model output will include the various state variables
at each time step.

For example, if discharge is in m^3/s, then the default \code{dt = 86400} corresponds to a
time step of 86400 seconds, or one day. Setting nt = 100 will run the model for 100 days.
}
