% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{species}
\alias{species}
\title{Creates species}
\usage{
species(c_params, e_params, alpha, beta, r_use, r_trans = identity)
}
\arguments{
\item{c_params}{Colonisation niche parameters, including \code{location}, \verb{breadth, }scale\code{and}dry`}

\item{e_params}{Extinction niche parameters, including \code{scale} and \code{dry}}

\item{alpha}{Active dispersal ability}

\item{beta}{Passive dispersal ability}

\item{r_use}{Resource use scaling parameter, one per resource axis}

\item{r_trans}{A transformation function for converting resources into niche dimensions}
}
\value{
An S3 object of class 'species', which contains the following named elements:
\itemize{
\item \code{col}: The colonisation function, takes a state matrix R and returns a vector of
colonisation rates
\item \code{ext}: The extinction function
\item \code{par_c}: colonisation niche parameters
\item \code{par_e}: extinction niche parameters
\item \code{alpha}: Active dispersal ability
\item \code{beta}: Passive dispersal ability
\item \code{r_use}: Resource use rate per niche axis
\item \code{r_trans}: Resource-to-niche transformation function
}

Additionally, the following attributes:
\itemize{
\item \code{niche_max}: the maximum possible value of the fundamental niche
}
}
\description{
Creates species
}
\details{
Creates a species, which is defined by its niche parameters and its
colonisation/extinction functions.

Note that for gaussian functions in multiple variables, \code{scale} is always a scalar, but
\code{location} must be a location vector of length \code{nx}, and \code{width} must be a variance covariance
matrix with \code{dim = c(nx, nx)}.

The resource use function is set automatically by default, but can be set to an arbitrary
function; see \link[=ruf]{resource use functions} for details on how this function should behave.
}
\examples{
NULL
}
