% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_network.r
\name{reset_state}
\alias{reset_state}
\alias{reset_species}
\alias{state}
\alias{state.river_network}
\alias{boundary.river_network}
\alias{state<-.river_network}
\alias{boundary<-.river_network}
\alias{site_by_species.river_network}
\alias{boundary_species.river_network}
\alias{boundary_species<-.river_network}
\alias{site_by_species<-.river_network}
\title{Get/set the (resource) state of a river network}
\usage{
reset_state(x, value)

reset_species(x, value)

\method{state}{river_network}(x, history = FALSE)

\method{boundary}{river_network}(x)

\method{state}{river_network}(x) <- value

\method{boundary}{river_network}(x) <- value

\method{site_by_species}{river_network}(x, history = FALSE)

\method{boundary_species}{river_network}(x)

\method{boundary_species}{river_network}(x) <- value

\method{site_by_species}{river_network}(x) <- value
}
\arguments{
\item{x}{A river network}

\item{value}{The value to update the attribute with}

\item{history}{Logical; if TRUE, entire state history is returned}
}
\description{
Setter and getter methods for river network state variables
}
\details{
By default, setting state will save the current state in the state history, then update
current state to \code{value}. \code{reset_state} erases the state variable and sets a new one, and does
not check that the dimensions make sense. Other methods update state and do dimension checking.

\code{boundary()} returns the boundary condition of the river network, as a site by resource matrix.
\verb{boundary() <-} sets the boundary; it must be a site by resoure matrix in the same format as
state.
}
